pyproject.toml
[project]
package-mode = false
name = "website-2025"
version = "0.1.0"
description = ""
authors = [
    {name = "nekell3r",email = "pazhabskiy032007@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13.2,<4.0"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "pydantic-extra-types (>=2.10.3,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.39,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "phonenumbers (>=9.0.1,<10.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "starlette (>=0.46.2,<0.47.0)",
    "redis (>=6.0.0,<7.0.0)",
    "bcrypt (==3.2.2)",
    "password-strength (>=0.0.3.post2,<0.0.4)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "ruff (>=0.11.9,<0.12.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "yookassa (>=3.5.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-dotenv (>=0.5.2,<0.6.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
]

[tool.poetry]
name = "website-2025"
version = "0.1.0"
description = ""
authors = ["..."]
package-mode = false
[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
